@@ -1,7 +1,7 @@
 describe V1::SpecimenEventsController do
 
   it 'GET #show' do
-    expect { get :show, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { get :show, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
   it 'GET #index' do
     expect(:get => "api/v1/patients/1/specimen_events?surgical_event_id=msn_3344").to route_to(:controller => "v1/specimen_events",
@@ -11,7 +11,7 @@ describe V1::SpecimenEventsController do
   end
 
   it 'GET #index should return something' do
-    get :index, :patient_id => "1234"
+    get :index, params: {patient_id: '1234'}
     expect(response).to have_http_status(200)
     expect(response.body).to include("{\"tissue_specimens\":[]")
     expect(response.body).to include("\"blood_specimens\"")
@@ -58,7 +58,7 @@ describe V1::SpecimenEventsController do
      :last_modified => "2016-11-11"}]
     allow(Aws::S3::S3Reader).to receive(:get_file_set).and_return(file_set)
 
-    get :index, :patient_id => "3366"
+    get :index, params: {patient_id: "3366"}
     expect(response).to have_http_status(200)
 
     events = JSON.parse(response.body).deep_symbolize_keys
@@ -70,7 +70,7 @@ describe V1::SpecimenEventsController do
     allow(NciMatchPatientModels::Specimen).to receive(:scan).and_return([{:surgical_event_id => "1234", :type => "TISSUE", :specimen_shipments => []}])
     allow(NciMatchPatientModels::Shipment).to receive(:scan).and_return([:molecular_id => "msn-1234"])
     allow(NciMatchPatientModels::VariantReport).to receive(:scan).and_return([])
-    get :index, :patient_id => "1234"
+    get :index, params: {patient_id: '1234'}
     expect(response).to have_http_status(200)
     expect(response.body).not_to be_blank
   end
@@ -79,21 +79,21 @@ describe V1::SpecimenEventsController do
     allow(NciMatchPatientModels::Specimen).to receive(:scan).and_return([{:surgical_event_id => "1234", :type => "BLOOD", :specimen_shipments => []}])
     allow(NciMatchPatientModels::Shipment).to receive(:scan).and_return([:molecular_id => "msn-1234"])
     allow(NciMatchPatientModels::VariantReport).to receive(:scan).and_return([])
-    get :index, :patient_id => "1234"
+    get :index, params: {patient_id: '1234'}
     expect(response).to have_http_status(200)
     expect(response.body).to include("specimen_shipments\":[]")
   end
 
   it 'POST #create' do
-    expect { post :create, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { post :create, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#update should throw an route error' do
-    expect { patch :update, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { patch :update, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#delete should throw an route error' do
-    expect { delete :destroy, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { delete :destroy, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
 end
