@@ -17,26 +17,26 @@ describe V1::SpecimensController do
 
   it 'GET #show will return something' do
     allow(NciMatchPatientModels::Specimen).to receive(:scan).and_return({})
-    get :show, patient_id: '123_test', id: "msn-123"
+    get :show, params: {patient_id: '123_test', id: "msn-123"}
     expect(response.body).to eq("")
   end
 
   it 'GET #index will return something' do
     allow(NciMatchPatientModels::Specimen).to receive(:scan).and_return({})
-    get :index, patient_id: '123_test'
+    get :index, params: {patient_id: '123_test'}
     expect(response.body).to eq("[]")
   end
 
   it 'POST #create' do
-    expect { post :create, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { post :create, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#update should throw an route error' do
-    expect { patch :update, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { patch :update, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#delete should throw an route error' do
-    expect { delete :destroy, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { delete :destroy, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
 end
