@@ -39,22 +39,22 @@ describe V1::StatisticsController do
   end
 
   it 'GET #show' do
-    expect { post :show, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { post :show, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
   it 'GET #index' do
     expect { get :index }.to raise_error(ActionController::UrlGenerationError)
   end
 
   it 'POST #create' do
-    expect { post :create, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { post :create, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#update should throw an route error' do
-    expect { patch :update, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { patch :update, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#delete should throw an route error' do
-    expect { delete :destroy, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { delete :destroy, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
 end
