@@ -14,28 +14,28 @@ describe V1::VariantsController do
 
   it 'GET #show will return something' do
     allow(NciMatchPatientModels::Variant).to receive(:scan).and_return({})
-    get :show, id: '123_test'
+    get :show, params: {id: '123_test'}
     expect(response.body).to eq("")
   end
 
   it 'GET #index will return something' do
     allow(NciMatchPatientModels::Variant).to receive(:scan).and_return({})
-    get :index, id: '123_test'
+    get :index, params: {id: '123_test'}
     expect(response.body).to eq("[]")
   end
 
 
 
   it 'POST #create' do
-    expect { post :create, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { post :create, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#update should throw an route error' do
-    expect { patch :update, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { patch :update, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
   it '#delete should throw an route error' do
-    expect { delete :destroy, :id => 1}.to raise_error(ActionController::UrlGenerationError)
+    expect { delete :destroy, params: {id: 1}}.to raise_error(ActionController::UrlGenerationError)
   end
 
 end
