@@ -1,7 +1,8 @@
 require File.expand_path('../boot', __FILE__)
-require File.expand_path('../version', __FILE__)
 
 require "rails"
+require File.expand_path('../version', __FILE__)
+
 # Pick the frameworks you want:
 require "active_model/railtie"
 require "action_controller/railtie"
