@@ -1,4 +1,4 @@
-# config/puma.rb
+port 10240, '0.0.0.0'
 threads 8,32
 workers 3
 preload_app!
