@@ -22,12 +22,12 @@ describe V1::EventsController, :type => :controller do
 
   it 'POST #create success' do
     allow(PatientProcessor).to receive(:run_service).and_return("test")
-    post :create, '{"patient_id": "123", "molecular_id": "123-mol", "analysis_id": "123-ana", "surgical_event_id": "123-surg", "rna_file_name": "test.bam" }'
+    post :create, params: {"patient_id": "123", "molecular_id": "123-mol", "analysis_id": "123-ana", "surgical_event_id": "123-surg", "rna_file_name": "test.bam" }, as: :json
     expect(response.body).to include("successfully")
   end
 
   it 'POST #create success' do
-    post :create, '{"entity_id": "123"}'
+    post :create, params: {"entity_id": "123"}, as: :json
     expect(response.status).to eq(403)
   end
 
